lockd.py: some code cleanups
Make sure we exit the program when the main thread
breaks by marking threads with daemon=True.
Now we don't need to explicitly exit the main thread
since that's what happens anyway with unhandled exceptions.
This should also make the exitcode != 0 so that
systemd's Restart=on-failure will restart the script.

Lastly we don't need to print timestamps now
everything is logged to the journal (with a timestamp).